refactor: made TrainManager immutable, use cases manager test into correct package folder
